refactor and rename files
